requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-core
langchain-openai
langchain-community
faiss-cpu
tiktoken
python-dotenv
